pyproject.toml
[project]
name = "mcp-server-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.115.14",
    "fastmcp>=2.10.1",
    "google-api-python-client>=2.173.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "langchain-google-genai>=2.1.5",
    "langchain-mcp-adapters>=0.1.7",
    "langgraph>=0.4.8",
    "uvicorn[standard]>=0.34.3",
]
